+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/ir/TracingTest.kt
@@ -52,6 +52,7 @@ class TracingTest : MetroCompilerTest() {
           ExampleGraph,Build component DAG
           ExampleGraph,Topo sort component DAG
           ExampleGraph,Expand components
+          ExampleGraph,Compute deferred init order
           ExampleGraph,Topo sort
           ExampleGraph,Sort and validate
           ExampleGraph,Compute binding indices
@@ -106,6 +107,8 @@ class TracingTest : MetroCompilerTest() {
                             ◀ Topo sort component DAG (xx ms)
                             ▶ Expand components
                             ◀ Expand components (xx ms)
+                            ▶ Compute deferred init order
+                            ◀ Compute deferred init order (xx ms)
                           ◀ Topo sort (xx ms)
                         ◀ Sort and validate (xx ms)
                         ▶ Compute binding indices
@@ -188,6 +191,7 @@ class TracingTest : MetroCompilerTest() {
           ChildGraphImpl,Build component DAG
           ChildGraphImpl,Topo sort component DAG
           ChildGraphImpl,Expand components
+          ChildGraphImpl,Compute deferred init order
           ChildGraphImpl,Topo sort
           ChildGraphImpl,Sort and validate
           ChildGraphImpl,Compute binding indices
@@ -207,6 +211,7 @@ class TracingTest : MetroCompilerTest() {
           ExampleGraph,Build component DAG
           ExampleGraph,Topo sort component DAG
           ExampleGraph,Expand components
+          ExampleGraph,Compute deferred init order
           ExampleGraph,Topo sort
           ExampleGraph,Sort and validate
           ExampleGraph,Compute binding indices
@@ -270,6 +275,8 @@ class TracingTest : MetroCompilerTest() {
                             ◀ Topo sort component DAG (xx ms)
                             ▶ Expand components
                             ◀ Expand components (xx ms)
+                            ▶ Compute deferred init order
+                            ◀ Compute deferred init order (xx ms)
                           ◀ Topo sort (xx ms)
                         ◀ Sort and validate (xx ms)
                         ▶ Compute binding indices
@@ -307,6 +314,8 @@ class TracingTest : MetroCompilerTest() {
                             ◀ Topo sort component DAG (xx ms)
                             ▶ Expand components
                             ◀ Expand components (xx ms)
+                            ▶ Compute deferred init order
+                            ◀ Compute deferred init order (xx ms)
                           ◀ Topo sort (xx ms)
                         ◀ Sort and validate (xx ms)
                         ▶ Compute binding indices
@@ -389,6 +398,7 @@ class TracingTest : MetroCompilerTest() {
           ChildGraphImpl,Build component DAG
           ChildGraphImpl,Topo sort component DAG
           ChildGraphImpl,Expand components
+          ChildGraphImpl,Compute deferred init order
           ChildGraphImpl,Topo sort
           ChildGraphImpl,Sort and validate
           ChildGraphImpl,Compute binding indices
@@ -408,6 +418,7 @@ class TracingTest : MetroCompilerTest() {
           ExampleGraph,Build component DAG
           ExampleGraph,Topo sort component DAG
           ExampleGraph,Expand components
+          ExampleGraph,Compute deferred init order
           ExampleGraph,Topo sort
           ExampleGraph,Sort and validate
           ExampleGraph,Compute binding indices
@@ -471,6 +482,8 @@ class TracingTest : MetroCompilerTest() {
                             ◀ Topo sort component DAG (xx ms)
                             ▶ Expand components
                             ◀ Expand components (xx ms)
+                            ▶ Compute deferred init order
+                            ◀ Compute deferred init order (xx ms)
                           ◀ Topo sort (xx ms)
                         ◀ Sort and validate (xx ms)
                         ▶ Compute binding indices
@@ -508,6 +521,8 @@ class TracingTest : MetroCompilerTest() {
                             ◀ Topo sort component DAG (xx ms)
                             ▶ Expand components
                             ◀ Expand components (xx ms)
+                            ▶ Compute deferred init order
+                            ◀ Compute deferred init order (xx ms)
                           ◀ Topo sort (xx ms)
                         ◀ Sort and validate (xx ms)
                         ▶ Compute binding indices