+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/BaseIncrementalCompilationTest.kt
@@ -35,11 +35,11 @@ abstract class BaseIncrementalCompilationTest {
     val keysPopulated: Set<String> by lazy {
       reportsDir.resolve("keys-populated-$name.txt").readLines().toSet()
     }
-    val providerFieldKeys: Set<String> by lazy {
-      reportsDir.resolve("keys-providerFields-$name.txt").readLines().toSet()
+    val providerPropertyKeys: Set<String> by lazy {
+      reportsDir.resolve("keys-providerProperties-$name.txt").readLines().toSet()
     }
-    val scopedProviderFieldKeys: Set<String> by lazy {
-      reportsDir.resolve("keys-scopedProviderFields-$name.txt").readLines().toSet()
+    val scopedProviderPropertyKeys: Set<String> by lazy {
+      reportsDir.resolve("keys-scopedProviderProperties-$name.txt").readLines().toSet()
     }
   }